+++ src/main/java/com/jwebmp/webawesome/components/WaCluster.java
@@ -2,7 +2,7 @@ package com.jwebmp.webawesome.components;
 
 import com.jwebmp.core.base.html.DivSimple;
 
-public class WaCluster<J extends WaCluster<J>> extends DivSimple<J> implements GapCapable<J>, AlignVerticalCapable<J>, SplitCapable<J>, SpaceTokenCapable<J>
+public class WaCluster<J extends WaCluster<J>> extends DivSimple<J> implements GapCapable<J>, AlignVerticalCapable<J>, JustifyContentCapable<J>, SplitCapable<J>, SpaceTokenCapable<J>
 {
     public WaCluster()
     {