+++ src/main/java/com/jwebmp/webawesome/components/WaDiv.java
@@ -1,6 +1,7 @@
 package com.jwebmp.webawesome.components;
 
 import com.jwebmp.core.base.html.DivSimple;
+import com.jwebmp.core.base.servlets.enumarations.ComponentTypes;
 
 /**
  * A general-purpose styled div that implements all Web Awesome token-capable interfaces.
@@ -9,15 +10,20 @@ import com.jwebmp.core.base.html.DivSimple;
  */
 public class WaDiv<J extends WaDiv<J>> extends DivSimple<J>
         implements SpaceTokenCapable<J>, TypographyTokenCapable<J>, BorderTokenCapable<J>,
-        ShadowTokenCapable<J>, ColourCapable<J>, TransitionTokenCapable<J>
+        ShadowTokenCapable<J>, ColourCapable<J>, TransitionTokenCapable<J>,
+        AlignVerticalCapable<J>,JustifyContentCapable<J>
 {
     public WaDiv()
     {
     }
 
-    public WaDiv(String tag)
-    {
-        setTag(tag);
+    public WaDiv(ComponentTypes myComponent) {
+        super(myComponent);
     }
+
+    public WaDiv(String text) {
+        super(text);
+    }
+
 }