+++ src/main/java/com/jwebmp/webawesome/components/WaStack.java
@@ -2,7 +2,7 @@ package com.jwebmp.webawesome.components;
 
 import com.jwebmp.core.base.html.DivSimple;
 
-public class WaStack<J extends WaStack<J>> extends DivSimple<J> implements GapCapable<J>, AlignVerticalCapable<J>, SpaceTokenCapable<J>
+public class WaStack<J extends WaStack<J>> extends DivSimple<J> implements GapCapable<J>, AlignVerticalCapable<J>, JustifyContentCapable<J>, SpaceTokenCapable<J>
 {
     public WaStack()
     {